Rework flowfield comms for server compatibility
