+++ fabric/gradle.properties
@@ -1,2 +1,2 @@
-loom.platform=neoforge
+loom.platform=fabric
 architectury_version=13.0.8
