+++ fabric/src/main/resources/fabric.mod.json
@@ -21,7 +21,7 @@
       "raccoonman.reterraforged.fabric.RTFFabric"
     ],
     "client": [
-      "raccoonman.reterraforged.fabric.client.FabricFlowFieldDebugRenderer"
+      "raccoonman.reterraforged.fabric.RTFFabricClient"
     ],
     "fabric-datagen": [
       "raccoonman.reterraforged.fabric.RTFFabric"
